+++ ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: mwubneh <mwubneh@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 15:07:28 by mwubneh           #+#    #+#             */
-/*   Updated: 2022/11/12 14:32:35 by mwubneh          ###   ########lyon.fr   */
+/*   Updated: 2022/11/12 16:43:41 by mwubneh          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (i <= n && s1 && s2 && s1[i] == s2[i])
